+++ WalletWasabi.Gui/Controls/WalletExplorer/BuildTabViewModel.cs
@@ -1,9 +1,7 @@
 using AvalonStudio.Extensibility;
 using AvalonStudio.Shell;
-using System;
+using NBitcoin;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 using WalletWasabi.Blockchain.TransactionBuilding;
 using WalletWasabi.Gui.Helpers;
@@ -13,33 +11,31 @@ namespace WalletWasabi.Gui.Controls.WalletExplorer
 {
 	public class BuildTabViewModel : SendControlViewModel
 	{
-		public BuildTabViewModel(Wallet wallet) 
-			: base(wallet, "Build Transaction", false)
+		public BuildTabViewModel(Wallet wallet)
+			: base(wallet, "Build Transaction")
 		{
 		}
 
 		public override string DoButtonText => "Build Transaction";
 		public override string DoingButtonText => "Building Transaction...";
 
-		protected override Task DoAfterBuildTransaction(BuildTransactionResult result)
+		public string PayjoinEndPoint => null;
+
+		protected override async Task BuildTransaction(string password,
+					PaymentIntent payments,
+			FeeStrategy feeStrategy,
+			bool allowUnconfirmed = false,
+			IEnumerable<OutPoint> allowedInputs = null)
 		{
-			try
-			{
-				var txviewer = new TransactionViewerViewModel();
-				IoC.Get<IShell>().AddDocument(txviewer);
-				IoC.Get<IShell>().Select(txviewer);
+			BuildTransactionResult result = await Task.Run(() => Wallet.BuildTransaction(Password, payments, feeStrategy, allowUnconfirmed: true, allowedInputs: allowedInputs));
 
-				txviewer.Update(result);
+			var txviewer = new TransactionViewerViewModel();
+			IoC.Get<IShell>().AddDocument(txviewer);
+			IoC.Get<IShell>().Select(txviewer);
 
-				ResetUi();
-
-				NotificationHelpers.Success("Transaction was built.");
-			}
-			catch (Exception ex)
-			{
-				return Task.FromException(ex);
-			}
-			return Task.CompletedTask;
+			txviewer.Update(result);
+			ResetUi();
+			NotificationHelpers.Success("Transaction was built.");
 		}
 	}
 }
