+++ WalletWasabi.Gui/Controls/WalletExplorer/SendTabViewModel.cs
@@ -1,6 +1,8 @@
 using NBitcoin;
 using NBitcoin.Payment;
+using ReactiveUI;
 using System;
+using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using WalletWasabi.Blockchain.Analysis.Clustering;
@@ -9,24 +11,42 @@ using WalletWasabi.Blockchain.Transactions;
 using WalletWasabi.Gui.Helpers;
 using WalletWasabi.Gui.Models.StatusBarStatuses;
 using WalletWasabi.Gui.ViewModels;
+using WalletWasabi.Gui.ViewModels.Validation;
 using WalletWasabi.Hwi;
 using WalletWasabi.Hwi.Exceptions;
+using WalletWasabi.Models;
 using WalletWasabi.Wallets;
+using WalletWasabi.WebClients.PayJoin;
 
 namespace WalletWasabi.Gui.Controls.WalletExplorer
 {
 	public class SendTabViewModel : SendControlViewModel
 	{
+		private string _payjoinEndPoint;
+
 		public SendTabViewModel(Wallet wallet)
-			: base(wallet, "Send", true)
+			: base(wallet, "Send")
 		{
 		}
 
 		public override string DoButtonText => "Send Transaction";
 		public override string DoingButtonText => "Sending Transaction...";
 
-		protected override async Task DoAfterBuildTransaction(BuildTransactionResult result)
+		[ValidateMethod(nameof(ValidatePayjoinEndPoint))]
+		public string PayjoinEndPoint
 		{
+			get => _payjoinEndPoint;
+			set => this.RaiseAndSetIfChanged(ref _payjoinEndPoint, value);
+		}
+
+		protected override async Task BuildTransaction(string password,
+			PaymentIntent payments,
+			FeeStrategy feeStrategy,
+			bool allowUnconfirmed = false,
+			IEnumerable<OutPoint> allowedInputs = null)
+		{
+			BuildTransactionResult result = await Task.Run(() => Wallet.BuildTransaction(Password, payments, feeStrategy, allowUnconfirmed: true, allowedInputs: allowedInputs, GetPayjoinClient()));
+
 			MainWindowViewModel.Instance.StatusBar.TryAddStatus(StatusType.SigningTransaction);
 			SmartTransaction signedTransaction = result.Transaction;
 
@@ -64,6 +84,26 @@ namespace WalletWasabi.Gui.Controls.WalletExplorer
 			ResetUi();
 		}
 
+		private IPayjoinClient GetPayjoinClient()
+		{
+			if (PayjoinEndPoint is { })
+			{
+				var payjoinEndPointUri = new Uri(PayjoinEndPoint);
+				return new PayjoinClient(payjoinEndPointUri, Global.TorManager.TorSocks5EndPoint);
+			}
+
+			return new NullPayjoinClient();
+		}
+
+		public ErrorDescriptors ValidatePayjoinEndPoint()
+		{
+			if (string.IsNullOrWhiteSpace(PayjoinEndPoint) || Uri.IsWellFormedUriString(PayjoinEndPoint, UriKind.Absolute))
+			{
+				return ErrorDescriptors.Empty;
+			}
+			return new ErrorDescriptors(new ErrorDescriptor(ErrorSeverity.Error, "Invalid url."));
+		}
+
 		protected override void OnAddressPaste(BitcoinUrlBuilder url)
 		{
 			base.OnAddressPaste(url);
@@ -72,6 +112,12 @@ namespace WalletWasabi.Gui.Controls.WalletExplorer
 						   || url.UnknowParameters.TryGetValue("pj", out endPoint)
 				? endPoint
 				: null;
+		}
+
+		protected override void ResetUi()
+		{
+			base.ResetUi();
+			PayjoinEndPoint = "";
 		}
 	}
 }
